Adding casts to help portability

+++ samples/MegaDrone/src/main/cpp/native-lib.cpp
@@ -48,7 +48,7 @@ JNIEXPORT jlong JNICALL
 Java_com_example_oboe_megadrone_MainActivity_startEngine(JNIEnv *env, jobject /*unused*/,
                                                          jintArray jCpuIds) {
     std::vector<int> cpuIds = convertJavaArrayToVector(env, jCpuIds);
-    LOGD("cpu ids size: %u", cpuIds.size());
+    LOGD("cpu ids size: %d", static_cast<int>(cpuIds.size()));
     MegaDroneEngine  *engine = new MegaDroneEngine(std::move(cpuIds));
     LOGD("Engine Started");
     return reinterpret_cast<jlong>(engine);
